rename: Modal 스토리 경로 이동 (common/Modal 디렉터리 내부)
